+++ CHANGELOG.md
@@ -22,4 +22,6 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Maintenance
 
+- Fix failed unit test due to React 18 upgrade ([#588](https://github.com/opensearch-project/dashboards-assistant/pull/642))
+
 ### Refactoring